Put together all tests that modify HeroesServer application state
